kafka_implementation/requirements.txt
pandas
psycopg2-binary
SQLAlchemy
requests
confluent-kafka